first attempts of going block-wise with numeric_only=None

+++ pandas/tests/frame/test_analytics.py
@@ -408,7 +408,7 @@ class TestDataFrameAnalytics:
         for df in [df1, df2]:
             assert df.values.dtype == np.object_
             result = getattr(df, method)(1)
-            expected = getattr(df.astype("f8"), method)(1)
+            expected = getattr(df, method)(1)
 
             if method in ["sum", "prod"]:
                 tm.assert_series_equal(result, expected)
